With excel data store
In this example, the events data is persisted using an excel file. the
PS script would read the contents of this excel file and make it
available to the end user in the form of a nice web UI.
I bought this timeline slider UI from a third party site.
The port opened is 8081.